ai-assistant/requirements.txt
vosk
sounddevice
pydub
numpy
llama-cpp-python
torch